requirements.txt
arrow==1.1.1
asgiref==3.4.1
binaryornot==0.4.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
cookiejar==0.0.3
Django==3.2.5
django-ckeditor==6.1.0
django-colorfield==0.4.2
django-comments-dab==2.6.1
django-crispy-forms==1.12.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-recaptcha==2.0.6
djangorestframework==3.12.4
idna==3.2
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
pager==3.3
Pillow==8.3.1
poyo==0.5.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-decouple==3.4
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
suds-py3==1.4.4.1
text-unidecode==1.3
urllib3==1.26.6
